+++ src/main/java/com/openggf/level/objects/DefaultObjectServices.java
@@ -508,8 +508,8 @@ public class DefaultObjectServices implements ObjectServices {
     }
 
     @Override
-    public void requestSpecialStageFromCheckpoint() {
-        lm().requestSpecialStageFromCheckpoint();
+    public void advanceToSpecialStageEntryRoutine() {
+        lm().advanceToSpecialStageEntryRoutine();
     }
 
     @Override